Feat: ledger need incometable

+++ src/main/java/com/kkh/jangbogimanager/ledger/dto/LedgerDto.java
@@ -10,8 +10,8 @@ public class LedgerDto {
 	private String id;
 	private String name;
 	private String invitation;
-	private int income;
-	private int expenses;
+	private Long income;
+	private Long expenses;
 	private Timestamp createAt;
 	private Timestamp updateAt;
 	private Member member;
